MB-68266: Fix narrowing warnings in utilities histogram printer
Fix narrowing conversion warnings in TimingHistogramPrinter:

 1. Change countFieldWidth() return type from int to size_t:
    - fmt::formatted_size() returns size_t
    - Returning it as int causes implicit narrowing conversion
    - size_t is the correct type for field width calculations

 2. Update caller in dump() to use const auto for numberOfSpaces:
    - Automatically deduces size_t from countFieldWidth()
    - Eliminates unnecessary narrowing

 3. Remove empty CMake conditional:
    - Delete unused "if Linux" block in utilities/CMakeLists.txt

These changes eliminate -Werror=shorten-64-to-32 warnings on 64-bit
platforms where size_t is wider than int.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/244294
Tested-by: Build Bot <build@couchbase.com>

+++ utilities/timing_histogram_printer.h
@@ -32,7 +32,7 @@ protected:
               double percentile);
 
     // Calculation for padding around the count in each histogram bucket
-    int countFieldWidth() const;
+    size_t countFieldWidth() const;
 
     // Calculation for histogram size rendering - how wide should the
     // ASCII bar be for the count of samples.
